configure restTemplates with SimpleClientHttpRequestFactory (#155)
With springboot 3.4.9 migration we get hangs with ReactorClientHttpRequestFactory (that is picked up automatically if spring-cloud is in the classpath with this version of springboot). We do this change to return to the SimpleClientHttpRequestFactory that was used to configure the RestTemplateBuilder using runtime configuration --spring.http.client.factory=simple
This absolutely requires that our servers use the spring RestTemplateBuilder, they can't new RestTemplateBuilder() themselves.

Signed-off-by: Abdelsalem <abdelsalem.hedhili@rte-france.com>